adding resolver, template_handler and renderer
